requirements.txt
flask
flask-cors
joblib
numpy
gunicorn
line-bot-sdk